Transaction - 1. View Bill 2. post Transaction 3.  search by date & customer Name

+++ InventoryManagement/Models/Transaction.cs
@@ -10,7 +10,7 @@ namespace InventoryManagement.Models
     {
         public int ID { get; set; }
 
-        public int TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
 
         public string ProductIDs { get; set; }
